run cookiecutter, add one crate (mc-bomb-types), mobilecoin submodule
This runs cookiecutter to set up github actions CI in this repo.

It adds a `mobilecoin.git` submodule, which is needed for lots of
reasons right now.

It also adds the simplest crate that we
can, the `mc-bomb-types` crate, which is roughly analogous to
`mc-fog-types`. This allows us to see that CI is working in a basic
sense.

Several key files are created as symlinks into the mobilecoin repo,
so for instance, `rust-toolchain`, `mob`, `.mobconf`, `hooks` and
`tools/install_git_settings.sh` are all symlinks into the mobilecoin
submodule, so that we don't end up forking them and follow the
mainline versions of these.